requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.31.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
spacy>=3.7.0
scikit-learn>=1.4.0
pydantic>=2.6.0
python-multipart>=0.0.9
requests>=2.31.0
pypdf>=4.2.0
numpy>=1.26.0
typing-extensions>=4.9.0
